fix(registry): harden previews and artifact projections
