+++ packages/registry/src/content-builder.js
@@ -1,7 +1,5 @@
 import path from "node:path";
 
-import matter from "gray-matter";
-
 import {
   deriveSeoFields,
   extractCodeBlocks,
@@ -12,20 +10,7 @@ import {
   normalizeBody,
 } from "./content-schema.js";
 import { buildBrandAssetMetadata } from "./brand-assets.js";
-
-const UNSAFE_FRONTMATTER_LANGUAGE_ERROR =
-  "Executable JavaScript frontmatter is not allowed in registry content";
-
-// Reject executable `---js` frontmatter instead of letting gray-matter's
-// JavaScript engine run untrusted submitted code during the content index
-// build. Mirrors the guard added for sibling parsers in #612. See #620.
-const SAFE_MATTER_OPTIONS = {
-  engines: {
-    javascript() {
-      throw new Error(UNSAFE_FRONTMATTER_LANGUAGE_ERROR);
-    },
-  },
-};
+import { parseSafeFrontmatter } from "./frontmatter.js";
 
 export const DEFAULT_DIRECTORY_REPO_URL =
   "https://github.com/JSONbored/awesome-claude";
@@ -224,7 +209,7 @@ export function buildContentEntryFromMdx(params) {
     contentUpdatedAt,
     getLocalDownloadSha256 = () => null,
   } = params;
-  const { data, content } = matter(source, SAFE_MATTER_OPTIONS);
+  const { data, content } = parseSafeFrontmatter(source);
   const body = normalizeBody(content, category);
   const headings = extractHeadings(body);
   const codeBlocks = extractCodeBlocks(body);
